Scripts to build the IPMs
These scripts use the estimates from the hierarchical models from JAGS and build IPMs with package ipmr. Then, some life history traits and demographic resilience components are calculated. Finally, a brute force sensitivity analysis is done to see which parameters in the IPM are more important for the resilience components.